tests/wpt/tests/websockets/interfaces/WebSocket/url/004.html

<!doctype html>
<title>WebSockets: 'URL'</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  assert_equals(ws.URL, undefined);
  assert_equals('URL' in ws, false);
  assert_equals(WebSocket.prototype.URL, undefined);
  assert_equals('URL' in WebSocket.prototype, false);
});
</script>
